remove components from AnyCloud MQTGT Client example that are not relevant for NanoDrone II
Publish task does not have to deal with the user button
Subscribe task optional, and does not haver to deal with user LED
Optional handling of lost connection now uses nanodrone topic
README reflects that the project is based on the AnyCloud example, but
subjects not relevant o this firmware removed, and UART and message
queue info provided. Wiring of the UART RX documented with serial
settings.
Images used in the REDME that are not relevant for this project removed.